requirements.txt
asgiref==3.4.1
attrs==21.4.0
blinker==1.4
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cryptography==3.4.8
Flask==2.0.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
imageio==2.13.5
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
kaitaistruct==0.9
ldap3==2.9.1
MarkupSafe==2.0.1
mitmproxy==7.0.4
msgpack==1.0.3
networkx==2.6.3
numpy==1.22.0
packaging==21.3
passlib==1.7.4
Pillow==9.0.0
pluggy==1.0.0
protobuf==3.18.1
publicsuffix2==2.20191221
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pytest==6.2.5
PyWavelets==1.2.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-image==0.19.1
scipy==1.7.3
six==1.16.0
sortedcontainers==2.4.0
tifffile==2021.11.2
toml==0.10.2
tornado==6.1
urwid==2.1.2
Werkzeug==2.0.2
wsproto==1.0.0
zstandard==0.15.2
